feat: size param for UserGroup
