--- src/harmony/UserGroup/UserGroup.stories.ts
@@ -1,24 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-
-import { UserGroup } from './UserGroup';
-
-const meta: Meta<typeof UserGroup> = {
-    title: '@harmony/UserGroup',
-    component: UserGroup,
-};
-
-export default meta;
-
-export const Default: StoryObj<typeof UserGroup> = {
-    args: {
-        users: [
-            { name: 'User 1', email: 'user1@example.com' },
-            { name: 'User 2', email: 'user2@example.com' },
-            { name: 'User 3', email: 'user3@example.com' },
-            { name: 'User 4', email: 'user4@example.com' },
-            { name: 'User 5', email: 'user5@example.com' },
-            { name: 'User 6', email: 'user6@example.com' },
-        ],
-        showCount: 4,
-    },
-};
